Make mprPutIntToBuf take an int64 Fix http.c: upload to handle huge files Fix http setting Host for ":4100" urls

+++ src/include/mpr.h
@@ -858,7 +858,7 @@ extern int mprPutBlockToBuf(MprBuf *buf, cchar *ptr, int size);
  *  @returns Zero if successful and otherwise a negative error code 
  *  @ingroup MprBuf
  */
-extern int mprPutIntToBuf(MprBuf *buf, int i);
+extern int mprPutIntToBuf(MprBuf *buf, int64 i);
 
 /**
  *  Put a string to the buffer.
